requirements.txt
Django
pgmpy
pandas
djangorestframework
wrapt
djangorestframework-recursive